SQLStuff has been split into 5 classes, updated multiple pages to reflect the changes.
