+++ FunWebsiteThing.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Content Include="SQLStuff.cs" />
+    <None Remove="database.db-journal" />
   </ItemGroup>
 
   <ItemGroup>
