+++ Pages/Comments.cshtml.cs
@@ -23,16 +23,16 @@ namespace FunWebsiteThing.Pages
         {
             CommentSection = int.TryParse(Request.Form["CS"], out int cs) ? cs : 0;
             string username = HttpContext.Session.GetString("Username") ?? "Anonymous";
-            await SQLStuff.AddComment(Comment, username, CommentSection);
+            await SQL.Comments.AddComment(Comment, username, CommentSection);
         }
 
         public async Task<IActionResult> OnPostDelete(int? commentid)
         {
             CommentSection = int.TryParse(Request.Form["CS"], out int cs) ? cs : 0;
             Logger.Write("CommentSection is " + CommentSection + ", TryParse is " + cs);
-            if (HttpContext.Session.GetInt32("IsAdmin") == 1)
+            if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId")))
             {
-                await SQLStuff.DeleteComment(commentid);
+                await SQL.Comments.DeleteComment(commentid);
             }
 
             return RedirectToPage("/Comments", new { cs = CommentSection });
