+++ Pages/Settings.cshtml
@@ -43,7 +43,7 @@
             }
         </script>
         <button onclick="GeneratePass()">Generate Pass</button>
-        if(HttpContext.Session.GetInt32("IsAdmin") == 1)
+        if (HttpContext.Session.GetInt32("IsAdmin") == 1 && SQL.Admin.IsAdmin(HttpContext.Session.GetInt32("UserId")))
         {
             <a href="/Admin">Admin Page</a>
         }
