+++ Pages/Update.cshtml
@@ -3,7 +3,7 @@
 @using System.Text.RegularExpressions;
 @model UpdateModel
 @{
-    ViewData["Title"] = "Updating " + SQLStuff.GetUsername(Model.Id);
+    ViewData["Title"] = "Updating " + SQL.Accounts.GetUsername(Model.Id);
 }
 
 @{
