--- SQL/SecurityQuestion.cs
@@ -1,6 +1,0 @@
-﻿namespace FunWebsiteThing.SQL
-{
-    public class SecurityQuestion
-    {
-    }
-}
